@@ -18,28 +18,28 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  cached_network_image: ^2.3.3
-  cloud_firestore: ^0.14.0+2
-  dropdown_search: ^0.4.4
-  firebase_auth: ^0.18.0+1
-  firebase_core: ^0.5.0
-  firebase_messaging: ^7.0.3
-  firebase_storage: ^4.0.0
-  flare_flutter: ^2.0.6
-  google_fonts: ^1.1.0
-  image_picker: ^0.6.7+7
+  cached_network_image: ^3.0.0
+  cloud_firestore: ^2.1.0
+  dropdown_search: ^0.6.1
+  firebase_auth: ^1.1.4
+  firebase_core: ^1.1.1
+  firebase_messaging: ^9.1.4
+  firebase_storage: ^8.0.6
+  flare_flutter: ^3.0.0
+  google_fonts: ^2.0.0
+  image_picker: ^0.7.5
 
-  url_launcher: ^5.7.2
+  url_launcher: ^6.0.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: ^0.8.1
+  flutter_launcher_icons: ^0.9.0
 
 flutter_icons:
   android: true
