@@ -11,22 +11,21 @@ class GetUserData {
           .doc(email)
           .get()
           .then((value) => {
-                UserProfileData.name = value.data()[AllKeys.nameKey],
-                UserProfileData.versityName =
-                    value.data()[AllKeys.versityNameKey],
-                UserProfileData.tmVersity = value.data()[AllKeys.tmVersityKey],
-                UserProfileData.phoneNum = value.data()[AllKeys.phnNumKey],
-                UserProfileData.address = value.data()[AllKeys.addressKey],
-                UserProfileData.admitted = value.data()[AllKeys.admittedKey],
-                UserProfileData.dept = value.data()[AllKeys.deptKey],
-                UserProfileData.email = value.data()[AllKeys.emailKey],
+                UserProfileData.name = value[AllKeys.nameKey],
+                UserProfileData.versityName = value[AllKeys.versityNameKey],
+                UserProfileData.tmVersity = value[AllKeys.tmVersityKey],
+                UserProfileData.phoneNum = value[AllKeys.phnNumKey],
+                UserProfileData.address = value[AllKeys.addressKey],
+                UserProfileData.admitted = value[AllKeys.admittedKey],
+                UserProfileData.dept = value[AllKeys.deptKey],
+                UserProfileData.email = value[AllKeys.emailKey],
                 UserProfileData.profilePicLink =
-                    value.data()[AllKeys.profilePicLinkKey],
+                    value[AllKeys.profilePicLinkKey],
                 UserProfileData.registrationNo =
-                    value.data()[AllKeys.registrationNoKey],
+                    value[AllKeys.registrationNoKey],
                 UserProfileData.uploadedBookNo =
-                    value.data()[AllKeys.upLoadedBookNoKey],
-                // UserProfileData.myBookList = value.data()[AllKeys.myBookListKey]
+                    value[AllKeys.upLoadedBookNoKey],
+                // UserProfileData.myBookList = value[AllKeys.myBookListKey]
               });
       if (UserProfileData.name == null) return 'empty';
       return 'done';
@@ -61,8 +60,8 @@ class GetBookData {
   static Future<List<BookData>> getRecent10Books() async {
     List<BookData> recentDataList = [];
 
-    QuerySnapshot querySnapshot = await FirebaseFirestore.instance
-        .collection(UserProfileData.tmVersity)
+    QuerySnapshot<Map<String,dynamic>> querySnapshot = await FirebaseFirestore.instance
+        .collection(UserProfileData.tmVersity!)
         .doc('AllBooks')
         .collection('AllBooks')
         .orderBy(AllKeys.bookTimeUploadKey, descending: true)
@@ -74,10 +73,11 @@ class GetBookData {
     return recentDataList;
   }
 
-  static Stream<QuerySnapshot> getRecentBookStream(int n, String req) {
+  static Stream<QuerySnapshot<Map<String, dynamic>>>? getRecentBookStream(
+      int n, String req) {
     if (UserProfileData.tmVersity != null)
       return FirebaseFirestore.instance
-          .collection(UserProfileData.tmVersity)
+          .collection(UserProfileData.tmVersity!)
           .doc('AllBooks')
           .collection(req)
           .orderBy(AllKeys.bookTimeUploadKey, descending: true)
@@ -87,7 +87,8 @@ class GetBookData {
     return null;
   }
 
-  static List<BookData> getBookDataObjFromQuerySnapshot(QuerySnapshot value) {
+  static List<BookData> getBookDataObjFromQuerySnapshot(
+      QuerySnapshot<Map<String, dynamic>> value) {
     List<BookData> recentDataList = [];
     value.docs.forEach((element) {
       BookData bookData = BookData();
@@ -99,46 +100,47 @@ class GetBookData {
     return recentDataList;
   }
 
-  static BookData getBookDataFromDocumentSnapshot(DocumentSnapshot element) {
+  static BookData getBookDataFromDocumentSnapshot(
+      DocumentSnapshot<Map<String, dynamic>> element) {
     BookData bookData = BookData();
 
     // print(element.data()[AllKeys.bookNameKey]);
-    bookData.bookName = element.data()[AllKeys.bookNameKey];
-    bookData.bookWriter = element.data()[AllKeys.bookWriterNameKey];
-    bookData.bookFor = element.data()[AllKeys.bookForKey];
-    bookData.bookDes = element.data()[AllKeys.bookDesKey];
-    bookData.bookImgLink = element.data()[AllKeys.bookImgKey];
-    bookData.bookPrice = element.data()[AllKeys.bookPriceKey];
-    bookData.bookEdition = element.data()[AllKeys.bookEditionKey];
-    bookData.bookTime = element.data()[AllKeys.bookTimeKey];
+    bookData.bookName = element.data()![AllKeys.bookNameKey];
+    bookData.bookWriter = element.data()![AllKeys.bookWriterNameKey];
+    bookData.bookFor = element.data()![AllKeys.bookForKey];
+    bookData.bookDes = element.data()![AllKeys.bookDesKey];
+    bookData.bookImgLink = element.data()![AllKeys.bookImgKey];
+    bookData.bookPrice = element.data()![AllKeys.bookPriceKey];
+    bookData.bookEdition = element.data()![AllKeys.bookEditionKey];
+    bookData.bookTime = element.data()![AllKeys.bookTimeKey];
 
-    bookData.bookTimeUpload = element.data()[AllKeys.bookTimeUploadKey];
+    bookData.bookTimeUpload = element.data()![AllKeys.bookTimeUploadKey];
     bookData.bookTimeUploadString =
-        element.data()[AllKeys.bookTimeUploadStringKey];
+        element.data()![AllKeys.bookTimeUploadStringKey];
 
-    bookData.bookUploaderName = element.data()[AllKeys.bookUploaderNameKey];
-    bookData.bookUploaderEmail = element.data()[AllKeys.bookUploaderEmailKey];
-    bookData.bookUploaderBatch = element.data()[AllKeys.bookUploaderBatchKey];
-    bookData.bookUploaderDept = element.data()[AllKeys.bookUploaderDeptKey];
-    bookData.bookUploaderImg = element.data()[AllKeys.bookUploaderImgKey];
+    bookData.bookUploaderName = element.data()![AllKeys.bookUploaderNameKey];
+    bookData.bookUploaderEmail = element.data()![AllKeys.bookUploaderEmailKey];
+    bookData.bookUploaderBatch = element.data()![AllKeys.bookUploaderBatchKey];
+    bookData.bookUploaderDept = element.data()![AllKeys.bookUploaderDeptKey];
+    bookData.bookUploaderImg = element.data()![AllKeys.bookUploaderImgKey];
     bookData.docId = element.reference.id;
     return bookData;
   }
 
-  static List<dynamic> bookNameList;
-  static Future<bool> getBookNameListFirebase() async {
+  static List<dynamic>? bookNameList;
+  static Future<bool?> getBookNameListFirebase() async {
     try {
       await FirebaseFirestore.instance
-          .collection(UserProfileData.tmVersity)
+          .collection(UserProfileData.tmVersity!)
           .doc('AllBooks')
           .get()
           .then((value) {
-        bookNameList = value.data()['FullNameArray'];
+        bookNameList = value['FullNameArray'];
       });
-      bookNameList.sort();
+      bookNameList!.sort();
       return true;
     } catch (e) {
-      bookNameList.add('No data');
+      bookNameList!.add('No data');
 
       return null;
     }
@@ -151,42 +153,42 @@ class GetBookData {
   //   return getBookDataFromDocumentSnapshot(dataRromRef);
   // }
 
-  static Future<QuerySnapshot> bookDataSearch(
-      String field, String search) async {
+  static Future<QuerySnapshot<Map<String, dynamic>>> bookDataSearch(
+      String field, String? search) async {
     return await FirebaseFirestore.instance
-        .collection(UserProfileData.tmVersity)
+        .collection(UserProfileData.tmVersity!)
         .doc('AllBooks')
         .collection('AllBooks')
         .where(field, isEqualTo: search)
         .get();
   }
 
-  static Stream<QuerySnapshot> bookDataSearchStream(
+  static Stream<QuerySnapshot<Map<String, dynamic>>> bookDataSearchStream(
       String field, String search) {
     return FirebaseFirestore.instance
-        .collection(UserProfileData.tmVersity)
+        .collection(UserProfileData.tmVersity!)
         .doc('AllBooks')
         .collection('AllBooks')
         .where(field, isEqualTo: search)
         .snapshots();
   }
 
-  static Future<QuerySnapshot> bookDataGrSearch(
+  static Future<QuerySnapshot<Map<String, dynamic>>> bookDataGrSearch(
       String field, String search) async {
     return await FirebaseFirestore.instance
-        .collection(UserProfileData.tmVersity)
+        .collection(UserProfileData.tmVersity!)
         .doc('AllBooks')
         .collection('AllBooks')
         .where(field, isGreaterThanOrEqualTo: search)
         .get();
   }
 
-  static bookDataDelete(String docId, String folder) async {
+  static bookDataDelete(String? docId, String folder) async {
     try {
       // print(docId);
       // print('\n');
       await FirebaseFirestore.instance
-          .collection(UserProfileData.tmVersity)
+          .collection(UserProfileData.tmVersity!)
           .doc('AllBooks')
           .collection(folder)
           .doc(docId)
@@ -201,8 +203,7 @@ class GetBookData {
 class StorageSettings {
   static deleteImage(String url) async {
     try {
-      StorageReference photoRef =
-          await FirebaseStorage.instance.getReferenceFromUrl(url);
+      Reference photoRef = FirebaseStorage.instance.refFromURL(url);
       await photoRef.delete();
     } catch (e) {
       print(e);
@@ -231,18 +232,19 @@ class StorageSettings {
 
 class Interactions {
   static DocumentReference firestoreRef = FirebaseFirestore.instance
-      .collection(UserProfileData.tmVersity)
+      .collection(UserProfileData.tmVersity!)
       .doc('Interactions');
 
   static writeMsg(String email, Map<String, dynamic> map) async {
     await firestoreRef.collection(email).add(map);
   }
 
-  static deleteDocWithId(String email, String id) async {
+  static deleteDocWithId(String email, String? id) async {
     await firestoreRef.collection(email).doc(id).delete();
   }
 
-  static Stream<QuerySnapshot> getMsgStream(String email) {
+  static Stream<QuerySnapshot<Map<String, dynamic>>> getMsgStream(
+      String email) {
     return firestoreRef
         .collection(email)
         .orderBy('SentKey', descending: true)
