@@ -4,9 +4,9 @@ import 'package:firebase_storage/firebase_storage.dart';
 import 'package:image_picker/image_picker.dart';
 
 class UploadIMG {
-  static PickedFile pickedImage;
-  static File imageUser;
-  static File imageBook;
+  static PickedFile? pickedImage;
+  static late File imageUser;
+  static late File imageBook;
 
   final _picker = ImagePicker();
 
@@ -16,7 +16,7 @@ class UploadIMG {
         imageQuality: 50,
         maxHeight: 720,
         maxWidth: 720);
-    imageUser = File(pickedImage.path);
+    imageUser = File(pickedImage!.path);
 
     // String dir = (await getApplicationDocumentsDirectory()).path;
     // String newPath = path.join(dir, 'case01wd03id01.jpg');
@@ -29,40 +29,40 @@ class UploadIMG {
         imageQuality: 50,
         maxHeight: 1200,
         maxWidth: 1200);
-    imageBook = File(pickedImage.path);
+    imageBook = File(pickedImage!.path);
   }
 
   uploadUserPic(String email, String id) async {
     // String id = UsableData.getMillisecondsId();
-    StorageReference userPic =
+    Reference userPic =
         FirebaseStorage.instance.ref().child('User').child(email).child(id);
 
-    await userPic.putFile(imageUser).onComplete;
+    await userPic.putFile(imageUser);
     return await userPic.getDownloadURL();
   }
 
   uploadBookPic(String email, String id) async {
-    StorageReference bookPic = FirebaseStorage.instance
+    Reference bookPic = FirebaseStorage.instance
         .ref()
         .child('Books')
         .child('Upload')
         .child(email)
         // .child(id)
         .child(id);
-    await bookPic.putFile(imageBook).onComplete;
+    await bookPic.putFile(imageBook);
     // print('upload Complete book');
     return await bookPic.getDownloadURL();
   }
 
   uploadRequstPic(String email, String id) async {
-    StorageReference bookPic = FirebaseStorage.instance
+    Reference bookPic = FirebaseStorage.instance
         .ref()
         .child('Books')
         .child('Request')
         .child(email)
         // .child(id)
         .child(id);
-    await bookPic.putFile(imageUser).onComplete;
+    await bookPic.putFile(imageUser);
     // print('upload Complete book');
     return await bookPic.getDownloadURL();
   }
