@@ -6,12 +6,12 @@ class FabMenu extends StatefulWidget {
   final Widget body;
   final List<FabMenuItem> items;
   final double blur;
-  final AnimatedIconData icon;
-  final Color fabColor;
-  final Color iconColor;
+  final AnimatedIconData? icon;
+  final Color? fabColor;
+  final Color? iconColor;
   FabMenu({
-    @required this.body,
-    @required this.items,
+    required this.body,
+    required this.items,
     this.blur: 5.0,
     this.icon,
     this.fabColor,
@@ -32,10 +32,10 @@ class _FabMenuState extends State<FabMenu> with TickerProviderStateMixin {
   Duration _duration = Duration(milliseconds: 500);
 
   /// Animation controller that animates the menu item
-  AnimationController _iconAnimationCtrl;
+  late AnimationController _iconAnimationCtrl;
 
   /// Animation that animates the menu item
-  Animation<double> _iconAnimationTween;
+  late Animation<double> _iconAnimationTween;
 
   @override
   void initState() {
@@ -175,8 +175,8 @@ class _MenuItemWidget extends StatelessWidget {
   final Function toggleMenu;
 
   _MenuItemWidget({
-    @required this.item,
-    @required this.toggleMenu,
+    required this.item,
+    required this.toggleMenu,
   });
 
   /// Closes the menu and calls the function for a particular menu item
@@ -238,18 +238,18 @@ class FabMenuItem {
   Function ontap;
 
   /// Background color for icon
-  Color color;
+  Color? color;
 
   /// Text color for label
-  Color labelColor;
+  Color? labelColor;
 
   /// Background color for label
-  Color labelBackgroundColor;
+  Color? labelBackgroundColor;
 
   FabMenuItem({
-    @required this.label,
-    @required this.ontap,
-    @required this.icon,
+    required this.label,
+    required this.ontap,
+    required this.icon,
     this.color,
     this.labelBackgroundColor,
     this.labelColor,
