@@ -5,7 +5,7 @@ import 'package:flutter/animation.dart';
 
 class SpinKitFadingCube extends StatefulWidget {
   const SpinKitFadingCube({
-    Key key,
+    Key? key,
     this.color,
     this.size = 50.0,
     this.itemBuilder,
@@ -15,14 +15,14 @@ class SpinKitFadingCube extends StatefulWidget {
             !(itemBuilder is IndexedWidgetBuilder && color is Color) &&
                 !(itemBuilder == null && color == null),
             'You should specify either a itemBuilder or a color'),
-        assert(size != null),
+     
         super(key: key);
 
-  final Color color;
+  final Color? color;
   final double size;
-  final IndexedWidgetBuilder itemBuilder;
+  final IndexedWidgetBuilder? itemBuilder;
   final Duration duration;
-  final AnimationController controller;
+  final AnimationController? controller;
 
   @override
   _SpinKitFadingCubeState createState() => _SpinKitFadingCubeState();
@@ -30,7 +30,7 @@ class SpinKitFadingCube extends StatefulWidget {
 
 class _SpinKitFadingCubeState extends State<SpinKitFadingCube>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
+  late AnimationController _controller;
 
   @override
   void initState() {
@@ -88,19 +88,19 @@ class _SpinKitFadingCubeState extends State<SpinKitFadingCube>
   }
 
   Widget _itemBuilder(int index) => widget.itemBuilder != null
-      ? widget.itemBuilder(context, index)
+      ? widget.itemBuilder!(context, index)
       : DecoratedBox(decoration: BoxDecoration(color: widget.color));
 }
 
 class DelayTween extends Tween<double> {
-  DelayTween({double begin, double end, this.delay})
+  DelayTween({double? begin, double? end, this.delay})
       : super(begin: begin, end: end);
 
-  final double delay;
+  final double? delay;
 
   @override
   double lerp(double t) =>
-      super.lerp((math.sin((t - delay) * 2 * math.pi) + 1) / 2);
+      super.lerp((math.sin((t - delay!) * 2 * math.pi) + 1) / 2);
 
   @override
   double evaluate(Animation<double> animation) => lerp(animation.value);
